OpenCV Integration
idk if this works cuz I can't import tensorflow and stuff from this file (Python 2 bad) but I think this is how you do it.
Make sure that the OpenCV_PlateFinder.scan_plate(imageName) line in CameraWindow() works cuz I can't test that and idk if that's how it works.